Add initial altcoin support
In preparation for supporting altcoins add a `Blockchain` type and
a bunch of structs and functions to assist in managing
blockchain-specific constants (currently just address prefixes).

Add support (prefixes) for Dogecoin and Litecoin.

Remove the `FromStr` implementation for `Address`. Add helper functions
for doing string conversion for bitcoin, litecoin, and dogecoin.

ltc: Add litecoin support

In order to support litecoin we need different address prefixes. We have
the infrastructure for this - add Litecoin.

As we did for dogecoin and bitcoin add a helper function to parse
strings into Litecoin addresses.